The issue with router and link was caused by the another main.jsx file that was conflicting with the one where we are working.Effectively we had 2 main.jsx.Deleted spare main.jsx file.
